+++ neural_compressor/pruner/pruner_legacy/pattern_lock.py
